requirements.txt
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///croot/attrs_1734533101012/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
datasets @ file:///croot/datasets_1716911606380/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill @ file:///croot/dill_1715094664823/work
einops==0.8.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///croot/filelock_1700591183607/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec @ file:///croot/fsspec_1714461537038/work
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
huggingface_hub @ file:///croot/huggingface_hub_1741284102044/work
idna @ file:///work/ci_py311/idna_1676822698822/work
ipython @ file:///croot/ipython_1704833016303/work
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib @ file:///croot/joblib_1718217211762/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
networkx @ file:///croot/networkx_1690561992265/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1709590491089/work/dist/pandas-2.2.1-cp311-cp311-linux_x86_64.whl#sha256=0a2793a31a0135a35735e1431d453a06186a3a7c607d9b441d9bd5f0fe4ded31
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1707233021655/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
propcache @ file:///croot/propcache_1744012718189/work
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp311-abi3-linux_x86_64.whl#sha256=8e447a0e89a4eb6aec01414ffb2b6e0032288e10ce1cf0a8129e73be2c0c34fb
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow @ file:///croot/pyarrow_1738229418212/work/python
Pygments @ file:///croot/pygments_1684279966437/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1707355572290/work
safetensors @ file:///croot/safetensors_1741361308036/work
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1743153247252/work/dist/scipy-1.15.2-cp311-cp311-linux_x86_64.whl#sha256=2ccc1711fbf2246e7fe52bafe286cce9fa222f297c222954db5caf8bb75e2e58
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///croot/sympy_1701397643339/work
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tokenizers @ file:///croot/tokenizers_1741370336077/work
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm @ file:///croot/tqdm_1738943501192/work
traitlets @ file:///work/ci_py311/traitlets_1676823305040/work
transformers @ file:///croot/transformers_1741705251010/work
triton==2.2.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1707770551213/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
xxhash @ file:///croot/python-xxhash_1737039434400/work
yarl @ file:///croot/yarl_1732546845924/work
